Create TravelingSalesMen
This is a simple non-elegant but functional TSP solver using permutations function for Python 2.7. Simply add your waypoints to goals as shown in the existing goals variable. The start can be modified but currently set for (0,0).

I wrote this to be put inside of another script with A* functionality so no plotting has been considered.